Use exec for direct ssh connect instead of syste0
The current usage of os.system doesn't interpret the ssh command
correctly. For exemple, the following command was failing. whereas it
works with standard ssh.

   ssh passhport@bastion mytarget "bash -c '( true ) && true'"

With os.system, it is not possible to choose the right separator. Either
quote or double quote can be wrong. os.exec take arg list or arg array
to avoid this problem.

Another good point is that the process become really ssh. So it more
transparent for the user of the bastion (less possible side effect).